+++ CHANGELOG.md
@@ -14,7 +14,9 @@ for details on updating existing applications using v1.x.y or v2.x.y.
 We are happy to announce the following GA libraries. Unless specifically noted,
 the APIs in these libraries are stable, and are ready for production use.
 
+- [Gemini Enterprise for Customer Experience API](google/cloud/ces/README.md)
 - [Vision AI API](/google/cloud/visionai/README.md)
+- [Workload Manager](/google/cloud/workloadmanager/README.md)]
 
 ## v3.2.0 - 2026-02